feat: Enable users to create songs individually
The data model has been modified to allow users to create and store songs individually in the database. A relationship has been added between the User model and the Song model to represent songs created by the user. This allows users Manage your own songs in the app.

Modified models:
- User: Added a UserSongs relationship to represent songs created by the user.
- Song: A user relationship has been added to establish the relationship between a song and the user who created it.